# ARROW-13534: [C++] Improve csv chunker