Add hints to help compiler deleting dead code per template augments.
Tested with clang-10, gcc-9.3, on x86 and Arm, observed ~20% uplift for
quoting=false benchmark. Probably due to better code locality.