@cyb70289 2021-08-03
`ChunkCSVEscapedBlock` test (quoting=false) improves significantly.
Below result is on xeon gold 5128. Similar result from Arm.

$ release/arrow-csv-parser-benchmark  --benchmark_filter="ChunkCSV.*"

**clang-10**
Before
```
ChunkCSVQuotedBlock         156626 ns       156625 ns         4483 bytes_per_second=1035.11M/s
ChunkCSVEscapedBlock        271061 ns       271063 ns         2585 bytes_per_second=562.925M/s
ChunkCSVNoNewlinesBlock        214 ns          214 ns      3278776 bytes_per_second=0/s
```
After
```
ChunkCSVQuotedBlock         156100 ns       156100 ns         4483 bytes_per_second=1038.6M/s
ChunkCSVEscapedBlock        204164 ns       204164 ns         3423 bytes_per_second=747.381M/s
ChunkCSVNoNewlinesBlock        214 ns          214 ns      3286563 bytes_per_second=0/s
```


**gcc-9.3**
Before
```
ChunkCSVQuotedBlock         218153 ns       218151 ns         3208 bytes_per_second=743.176M/s
ChunkCSVEscapedBlock        225251 ns       225251 ns         3107 bytes_per_second=677.414M/s
ChunkCSVNoNewlinesBlock        202 ns          202 ns      3467515 bytes_per_second=0/s
```
After
```
ChunkCSVQuotedBlock         222775 ns       222776 ns         3141 bytes_per_second=727.747M/s
ChunkCSVEscapedBlock        186540 ns       186541 ns         3751 bytes_per_second=817.984M/s
ChunkCSVNoNewlinesBlock        202 ns          202 ns      3469361 bytes_per_second=0/s
```