@pitrou 2021-08-03
Nice! Can the same technique be applied to the CSV parser?