@cyb70289 2021-08-03
Didn't investigate deep but it looks easy for compiler to figure out the dead code in CSV parser.
CSV chunker::ReadLine is "stateful", so it's harder for compiler.